!45 fix: 删除无用的函数和成员 * fix: 注释掉cout * fix: 使用emplace_back() * fix: 移除无用语句 * fix: 删除无用的函数和成员

+++ src/AI/Node.cpp
@@ -64,7 +64,7 @@ float Node::refreshAvgValue()
 
 float Node::getUCBValue(int Total)
 {
-    return AvgValue + sqrt((((log((float)Total)) * 2) / (float)Times));
+    return AvgValue + sqrt((log(Total) * 2) / Times);
 }
 
 Node *Node::expandUCTNode()
